+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - Show translation keys debug tool will now translate keys when Alt is held.
 - On 1.21 and above, support for Polymer's custom items.
+- Item Frames now show the description of the item inside, not their own descriptions.
 
 ## Fixed
 - Incorrect descriptions on some leaf blocks.